% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generatePGN.R
\name{generatePGN}
\alias{generatePGN}
\title{Protein-Gene Network Construction}
\usage{
generatePGN(glist, tftg, ppi, tftg_thre = NULL, ptf_thre = NULL,
  ppi_thre = NULL)
}
\arguments{
\item{glist}{The vector of genes in the same order of the genes in log2FC data. The length of GList should be the same as the number of rows in log2FC data.}

\item{tftg}{The matrix of TF-gene interactions. The first column is the list of TFs, and the second column is the list of genes regulated by the corresponding TFs. The third column is optional, and if present, the column should contain the (confidence) score for each interaction.}

\item{ppi}{The matrix of protein-protein interactions. Each row of the first two columns give the protein pairs with interactions. The third column is optional, and if present, the column should contain the (confidence) score for each interaction.}

\item{tftg_thre}{A threshold for TF-gene interactions. This variable is used only when the confidence score of TF-gene interactions are given in the matrix tftg. Any TF-gene interactions with confidence scores lower than the threshold will be excluded.}

\item{ptf_thre}{A threshold for protein-TF interaction. This variable is used only when the confidence score of protein-TF interactions are given in the matrix ppi.  Any protein-TF interactions with the scores lower than the threshold will be excluded.}

\item{ppi_thre}{A threshold for protein-protein interaction. This variable is used only when the confidence score of protein-protein interactions are given in the matrix ppi.  Any protein-protein interactions with the scores lower than the threshold will be excluded.}
}
\value{
\item{pgn}{The adjacency matrix of PGN}
}
\description{
Generate protein-gene regulatory network from gene regulatory network and protein protein interaction network
}
