% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/salmon.R
\name{salmon}
\alias{salmon}
\title{SALMON}
\usage{
salmon(lfc, slope = NULL, pgn, grplist, kfold = 10, par = FALSE,
  numCores = 4)
}
\arguments{
\item{lfc}{The numeric matrix or data.frame of log2FC data. Each row represents a gene and each column represents a sample.}

\item{slope}{The slope matrix from log2FC data. This matrix can be obtained using the function \code{generateSlope()}. If the data are not time-series, set slope to an empty matrix (i.e. \code{slope}=\code{NULL}).}

\item{pgn}{The adjacency matrix of the protein-gene regulation network. This matrix can be created using the function \code{generatePGN()}.}

\item{grplist}{The group index for protein scoring. This vector defines the samples for which the protein scores are computed. The length of this vector should be the same as the number of samples in the log2FC matrix. A single (aggregate) protein score is generated for samples with the same index. The group indices should be a consecutive integer starting from 1 to the number of groups.}

\item{kfold}{The number of folds used in the k-fold cross validation.}

\item{par}{A Boolean variable \code{TRUE} or \code{FALSE} indicating whether to use parallel computing. The default is \code{FALSE} (no parallel computation).}

\item{numCores}{The number of CPU cores to be used for parallel computing. This parameter is considered only if par is \code{TRUE}. The default is 4.}
}
\value{
a list of two matrices
\item{Pscore}{The matrix of protein scores. Each row corresponds to a gene following the same order as the one in the log2FC data, while each column corresponds to a group of samples as defined in the grplist.}
\item{A}{The matrix of PGN edges. The rows correspond to genes having at least one regulator based on the PGN (i.e. zeros for the others).}
}
\description{
The main function for SALMON for generating the protein scores for each drug treatment.
}
